% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{methods}
\name{addFile,simulation-method}
\alias{addFile,simulation-method}
\title{Run a QuantiNemo simulation}
\usage{
\S4method{addFile}{simulation}(object, file.name, file.content)
}
\arguments{
\item{file.name}{The name of the file that you want to add, as defined by QuantiNemo}

\item{file.content}{A variable of dataframe type containing the needed information}
}
\description{
Run a QuantiNemo simulation
}
\examples{
my_sim <- new("simulation")
my_sim <- setParameter(my_sim,"quanti_all",5)
alleles <- data.frame(locus = rep(1,5),allele = seq(1,5),value = seq(-2,2))
my_sim <-  addFile(my_sim, "quanti_allelic_file",alleles)
run(my_sim)
}
