% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{addFile}
\alias{addFile}
\title{Add an input file  to quantiNemo.}
\usage{
addFile(object, file.name, file.content)
}
\arguments{
\item{file.name}{The name of the file that you want to add, as defined by QuantiNemo}

\item{file.content}{A dataframe type containing the needed information. The name of the Column should be the one given by Quantinemo}
}
\description{
Add an input file  to quantiNemo.
}
\examples{
my_sim <- new("simulation")
my_sim <- setParameter(my_sim,"quanti_all",5)
alleles <- data.frame(locus = rep(1,5),allele = seq(1,5),value = seq(-2,2))
my_sim <-  addFile(my_sim, "quanti_allelic_file",alleles)
run(my_sim)
}
