% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\docType{methods}
\name{getCommand,simulation-method}
\alias{getCommand,simulation-method}
\title{get the command to be run}
\usage{
\S4method{getCommand}{simulation}(object)
}
\description{
get the command to be run
}
\examples{
my_simulation <- new("simulation")
system(getCommand(my_simulation))
}
