% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{getCommand}
\alias{getCommand}
\title{get the command to be run}
\usage{
getCommand(object)
}
\description{
get the command to be run
}
\examples{
my_simulation <- new("simulation")
system(getCommand(my_simulation))
}
