% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{getParameter}
\alias{getParameter}
\title{Get the value of a parameter if it exist or the default value otherwise}
\usage{
getParameter(object, parameter, default = 0)
}
\arguments{
\item{parameter}{the parameter which we want to extract}

\item{default}{The default value which will be return in case the parameter does not exist}
}
\description{
Get the value of a parameter if it exist or the default value otherwise
}
\examples{
my_sim <- new("simulation", parameters = list("generations" = 100)
get_parameter(my_sim, "generations") # return 100
get_parameter(my_sim, "patch_number") # return 0
get_parameter(my_sim, "patch_number", default = 1) # return 1
}
