% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{getPostInfo}
\alias{getPostInfo}
\title{Get the last part of a filename that changes through generation and replicate}
\usage{
getPostInfo(object, generation = -1, replicate = 0)
}
\arguments{
\item{generation}{The generation from which we want to load the data. A Negative number mean starting from the end.}

\item{replicate}{The replicate from which we want to load the data. A value of 0 mean that there is only one replicate.}
}
\description{
Get the last part of a filename that changes through generation and replicate
}
\examples{
my_sim <- new("simulation")
getPostInfo(my_sim, generation = 10, replicate = 1)
}
