% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{loadGeno}
\alias{loadGeno}
\title{Load the Genotype of a QuantiNemo simulation}
\usage{
loadGeno(object, generation = -1, replicate = 0)
}
\arguments{
\item{generation}{The generation from which we want to load the data. A Negative number mean starting from the end.}

\item{replicate}{The replicate from which we want to load the data. A value of 0 mean that there is only one replicate.}
}
\description{
Load the Genotype of a QuantiNemo simulation
}
\examples{
my_sim <- new("simulation")
my_sim <- setParameter(my_sim,"quanti_loci",1)
my_sim <- setParameter(my_sim,"quanti_save_genotype",1)
run(my_sim)
geno <- loadGeno(my_sim)
table(geno$V2)
}
