% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{methods}
\name{loadPheno,simulation-method}
\alias{loadPheno,simulation-method}
\title{Load the Phenotype of a QuantiNemo simulation}
\usage{
\S4method{loadPheno}{simulation}(object, generation = -1, replicate = 0)
}
\description{
Load the Phenotype of a QuantiNemo simulation
}
\examples{
my_sim <- new("simulation")
my_sim <- setParameter(my_sim,"quanti_loci",1)
my_sim <- setParameter(my_sim,"quanti_save_phenotype",1)
run(my_sim)
my_sim <- loadPheno(my_sim)
table(my_sim@Pheno$V2)
}
