% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{loadPheno}
\alias{loadPheno}
\title{Load the Phenotype of a QuantiNemo simulation}
\usage{
loadPheno(object, generation = -1, replicate = 0)
}
\arguments{
\item{generation}{The generation from which we want to load the data. A Negative number mean starting from the end.}

\item{replicate}{The replicate from which we want to load the data. A value of 0 mean that there is only one replicate.}
}
\description{
Load the Phenotype of a QuantiNemo simulation
}
\examples{
my_sim <- new("simulation")
my_sim <- setParameter(my_sim,"quanti_loci",1)
my_sim <- setParameter(my_sim,"quanti_save_phenotype",1)
run(my_sim)
pheno <- loadPheno(my_sim)
table(pheno$V2)
}
