% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{methods}
\name{loadStat,simulation-method}
\alias{loadStat,simulation-method}
\title{Load the statistic of a QuantiNemo simulation}
\usage{
\S4method{loadStat}{simulation}(object)
}
\description{
Load the statistic of a QuantiNemo simulation
}
\examples{
my_sim <- new("simulation")
my_sim <- setParameter(my_sim,"stat","adlt.ind")
run(my_sim)
my_sim <- loadStat(my_sim)
plot(my_sim@stat$adlt.ind)
}
