% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{loadStatPatch}
\alias{loadStatPatch}
\title{Load the statistic (mean) of a QuantiNemo simulation for each patch and return it as a matrix}
\usage{
loadStatPatch(object, stat.name)
}
\arguments{
\item{stat.name}{String representing the name of the statisitic to load}
}
\description{
Load the statistic (mean) of a QuantiNemo simulation for each patch and return it as a matrix
}
\examples{
parameters = list("generations" = 5,
                  "patch_capacity" = 100,
                  "patch_number" = 1000,
                  "stat" = "{adlt.nbInd_p}",
                  "patch_ini_size", "{seq(1,1000,1000)}")
my_sim.base = new("simulation", parameters = parameters)
run(my_sim, verbose =FALSE)
plot(loadStatPatch(my_sim, "adlt.nbInd_p")[1, ])
}
