% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{loadStatRep}
\alias{loadStatRep}
\title{Load the statistic for various replicate of a QuantiNemo simulation}
\usage{
loadStatRep(object)
}
\description{
Load the statistic for various replicate of a QuantiNemo simulation
}
\examples{
my_sim <- new("simulation")
my_sim <- setParameter(my_sim,"stat","adlt.ind")
run(my_sim)
stat.r <- loadStatRep(my_sim,"stat","adlt.ind")
for (i in 1:10){
  plot(stat.r$adlt.ind[stat.r$replicate==i])
 }
 @export
}
