% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/plot.R
\name{print.simulation}
\alias{print.simulation}
\alias{print.simulation}
\title{Display basic information about a quantiNemo simulation}
\usage{
\method{print}{simulation}(object, ...)

\method{print}{simulation}(object, ...)
}
\arguments{
\item{object}{The simulation to be plotted}

\item{file.name}{The name of the file that you want to add, as defined by QuantiNemo}

\item{file.content}{A variable of dataframe type containing the needed information}

\item{statistics}{A character or vector of character countaining the names 
of the statistics to be plotted}

\item{dispatch}{decide if the plot are plotted all on the same page or on various
page}
}
\description{
Display basic information about a quantiNemo simulation

This function allows to plot the main statistic from a simulation.
By default, it plot all available statistics on several plot in a single page
, but can alternatively plot the different statistics on different plot
or the same statistics for various replicate
}
\examples{
my_sim <- new("simulation", sim.name = "test72")
my_sim <- setParameter(my_sim,"quanti_all",5)
print(my_sim)
my_sim <- new("simulation", sim.name = "test_plot")
my_sim <- setParameter(my_sim,"stat","adlt.nb_individual")
run(my_sim)
plot(my_sim)
}
