\name{RQuantiNemo-package}
\alias{RQuantiNemo-package}
\alias{RQuantiNemo}
\docType{package}
\title{
\packageTitle{RQuantiNemo}
}
\description{
\packageDescription{RQuantiNemo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RQuantiNemo}
\packageIndices{RQuantiNemo}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RQuantiNemo}

Maintainer: \packageMaintainer{RQuantiNemo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
