% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{methods}
\name{run,simulation-method}
\alias{run,simulation-method}
\title{Run a QuantiNemo simulation}
\usage{
\S4method{run}{simulation}(object, verbose = TRUE)
}
\description{
Run a QuantiNemo simulation
}
\examples{
my_simulation <- new("simulation")
run(my_simulation)
}
