% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{run}
\alias{run}
\title{Run a QuantiNemo simulation}
\usage{
run(object, verbose = TRUE)
}
\arguments{
\item{verbose}{If we want to display information about the simulation and the standard output of QuantiNemo.}
}
\description{
Run a QuantiNemo simulation
}
\examples{
my_simulation <- new("simulation")
run(my_simulation)
}
