% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{methods}
\name{setParameter,simulation-method}
\alias{setParameter,simulation-method}
\title{Run a QuantiNemo simulation}
\usage{
\S4method{setParameter}{simulation}(object, name, value)
}
\arguments{
\item{name}{The name of the parameter that you want to set. It has to be a known parameter for QuantiNemo.}

\item{value}{The value of the parameter that you want to set. It can be a number if it's a simple parameter or a string for more complex parameter}
}
\description{
Run a QuantiNemo simulation
}
\examples{
my_sim <- new("simulation")
my_sim  <- setParameter(my_sim,"quanti_all",5)
run(my_simulation)
}
