% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{setParameter}
\alias{setParameter}
\title{Add a new parameter to a quantiNemo simulation.}
\usage{
setParameter(object, name, value)
}
\arguments{
\item{name}{The name of the parameter that you want to set. It has to be a known parameter for QuantiNemo.}

\item{value}{The value of the parameter that you want to set. It can be a number if it's a simple parameter or a string for more complex parameter}
}
\description{
Add a new parameter to a quantiNemo simulation.
}
\examples{
my_sim <- new("simulation")
my_sim  <- setParameter(my_sim,"patch_number",5)
my_sim  <- setParameter(my_sim,"patch_ini_size","{10 100 10 100 67}")
}
