% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\docType{methods}
\name{writeDataframeFiles,simulation-method}
\alias{writeDataframeFiles,simulation-method}
\title{Print all the input files which come from a data frame}
\usage{
\S4method{writeDataframeFiles}{simulation}(object)
}
\description{
Print all the input files which come from a data frame
}
\examples{
my_simulation <- new("simulation")
writeDataframeFile(my_simulation)
}
