% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{writeDataframeFiles}
\alias{writeDataframeFiles}
\title{Print all the input files which come from a data frame}
\usage{
writeDataframeFiles(object)
}
\description{
Print all the input files which come from a data frame
}
\examples{
my_simulation <- new("simulation")
writeDataframeFile(my_simulation)
}
