% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\docType{methods}
\name{writeGenoFile,simulation-method}
\alias{writeGenoFile,simulation-method}
\title{Print the genotype file}
\usage{
\S4method{writeGenoFile}{simulation}(object)
}
\description{
Print the genotype file
}
\examples{
my_simulation <- new("simulation")
writeMainFile(my_simulation)
}
