% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{writeGenoFile}
\alias{writeGenoFile}
\title{Print the genotype file}
\usage{
writeGenoFile(object)
}
\description{
Print the genotype file
}
\examples{
my_simulation <- new("simulation")
writeMainFile(my_simulation)
}
