% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\docType{methods}
\name{writeInput,simulation-method}
\alias{writeInput,simulation-method}
\title{Print all the input file for a QuantiNemo simulation}
\usage{
\S4method{writeInput}{simulation}(object)
}
\description{
Print all the input file for a QuantiNemo simulation
}
\examples{
my_simulation <- new("simulation")
writeInput(my_simulation)
}
