% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{writeInput}
\alias{writeInput}
\title{Print all the input file for a QuantiNemo simulation}
\usage{
writeInput(object, verbose = TRUE)
}
\description{
Print all the input file for a QuantiNemo simulation
}
\examples{
my_simulation <- new("simulation")
writeInput(my_simulation)
}
