% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\docType{methods}
\name{writeMainFile,simulation-method}
\alias{writeMainFile,simulation-method}
\title{Print the main input file for the QuantiNemo simulation}
\usage{
\S4method{writeMainFile}{simulation}(object)
}
\description{
Print the main input file for the QuantiNemo simulation
}
\examples{
my_simulation <- new("simulation")
writeMainFile(my_simulation)
}
