% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{writeMainFile}
\alias{writeMainFile}
\title{Print the main input file for the QuantiNemo simulation}
\usage{
writeMainFile(object, verbose = TRUE)
}
\description{
Print the main input file for the QuantiNemo simulation
}
\examples{
my_simulation <- new("simulation")
writeMainFile(my_simulation)
}
